Various tweaks to components. Add initial local storage service and a… (#13)
* Various tweaks to components. Add initial local storage service and add dark theme.

* Create title service and remove code from app component

Take all of the title changing logic out of the main app
component.ts file and put in it's own service. The title service
is fed in to app.component.ts through the constructor to register
changing of the title.

* Update about component with new structure

Update the about component, replacing 'hard coded' values with
configured objects instead. Update the layout for new data sources,
and update small about of css for about component.

* Remove unused imports in app.component.ts